src/ClientFirst/components/prepare_base_model.py
import pandas as pd
from pathlib import Path
from sklearn.preprocessing import LabelEncoder
from ClientFirst.utils.logger import logger
from ClientFirst.utils.common import save_joblib
from ClientFirst.entity.config_entity import PrepareBaseModelConfig


class PrepareBaseModel:
    def __init__(self, config: PrepareBaseModelConfig):
        self.config = config

    def prepare_encoders(self):
        """Prepare and save label encoders and categories"""
        try:
            # This is a placeholder - actual implementation depends on data
            logger.info("Prepare base model components initialized")
            logger.info(f"Artifacts will be saved to {self.config.root_dir}")
            
        except Exception as e:
            logger.error(f"Error preparing base model: {e}")
            raise e
