requirements.txt
# requirements.txt: This file lists the Python packages required for the ClientFirst-XAI project.
# Core ML Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
catboost>=1.2.0

# Deep Learning & XAI
shap>=0.42.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2

# Utilities
joblib>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
loguru>=0.7.0
typing-extensions>=4.8.0

# Configuration Management
python-box[all]>=7.1.0
ensure>=1.0.2
PyYAML>=6.0.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
flake8>=6.1.0
black>=23.10.0

# Email
aiosmtplib>=2.0.0